test for compliance with a strict CSP
Ember CLI Content Security Policy ships test support in it's upcoming v2
release. It could be used to test compliance with a strict CSP. If a
test violates the configured CSP it will fail. The addon comes with
reasonable defaults and does not require any configuration for this use
case.