Stop using ember-get-config
ember-get-config isn't necessary, configuration can be retrieved via the notification service's owner.

When running tests from the /tests URL after starting a development server with `ember serve`, the configuration returned by ember-get-config is stuck to development environment, whereas the configuration returned by the service's owner is from test environment.
This is specific to `ember serve`, it is not the case when using `ember test --server`.